backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
blinker==1.9.0
    # via flask
click==8.3.0
    # via flask
filelock==3.20.0
    # via torch
flask==3.1.2
    # via
    #   backend (pyproject.toml)
    #   flask-cors
flask-cors==6.0.1
    # via backend (pyproject.toml)
fsspec==2025.10.0
    # via torch
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   torch
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.3.4
    # via torchvision
pillow==12.0.0
    # via torchvision
setuptools==80.9.0
    # via torch
sympy==1.14.0
    # via torch
torch==2.9.0+cpu
    # via
    #   backend (pyproject.toml)
    #   torchvision
torchvision==0.24.0+cpu
    # via backend (pyproject.toml)
typing-extensions==4.15.0
    # via torch
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
